logmsg: use higher level nvtable interface in log_msg_get_value()
The function log_msg_get_value() is an absolute fast-path, potentially
called several times for every message received. That has probably been
the reason to pass LM_V_MAX as a constant as "num_static_entries" parameter,
so that the inline functions in nvtable.h can be optimized accordingly.

This has been broken at one point, so this argument went unused,
and we always used self->num_static_entries instead of the passed
in argument. I am not sure how much performance we could gain by restoring
the potential optimization, but instead of doing that I simply removed
the coupling, making the code simpler.

If this optimization will be a bottleneck in the future, we can still
restore it.

Signed-off-by: Balazs Scheidler <balazs.scheidler@balabit.com>