nvtable: add the concept of "unset" values
Unset values are either those that have not yet been set in the message or
those that have explicitly been unset.

Previously, unset values simply returned the empty string, making it impossible
to find out if a name-value was explicitly set to the empty string.

This patch introduces a two functions at the NVTable level:

  - nv_table_unset()
  - nv_table_get_value_if_set()

nv_table_get() will continue to return an empty string for unset values.

Signed-off-by: Balazs Scheidler <balazs.scheidler@balabit.com>