logmsg: introduce global state struct for a serialization/deserialization operation
The various bits that make up the LogMessage serialization (LogMsg,
NVTable, SData, tags, etc) are pretty tightly coupled, however state
only travels in function arguments (the message under serialization,
and the archive where we write/read bytes to/from).

This makes the code rigid, as it is difficult to introduce additional state.

For instance the version number being used for serialization is not
passed around, which we need in case we need to update formats.

This patch introduces this state and starts using it at the highest level
functions (e.g. log_msg_deserialize/serialize).

Signed-off-by: Balazs Scheidler <balazs.scheidler@balabit.com>