nvtable-serialize: introduce _fixup_entry() that will cover all fixup operations
This patch is the first in a series to refactor the code that iterated
over NVTable multiple times, to something that does the same with a single
iteration, and hopefully more readable code.

Basically the plan is to run nv_table_foreach_entry() with a callback that
does all the different update operations that are relevant to that entry.

This patch only focuses on RFC5424 structured data entries.

Signed-off-by: Balazs Scheidler <balazs.scheidler@balabit.com>