nvtable: pass NVDynValue to nv_table_foreach_entry() callbacks
nv_table_foreach_entry() is an internal access function to iterate over all
values within an NVTable.  In deserialization we would need access to the
NVDynValue as well as NVEntry in order to update that. This patch
adds this additional argument to the callback.

We do have a user-visible interface (nv_table_foreach) which does not
care about internals and whether a name-value pair is dynamic or not and
that is not affected by this change.

Signed-off-by: Balazs Scheidler <balazs.scheidler@balabit.com>