testutils: add iteration argument to stopwatch function
There's a handy stopwatch function in testutils that displays the amount
of time a code needed to execute.

This patch adds a parameter "iterations" to tell the function the
number of loops we were going through, so that it can report the
iterations per second in addition to the runtime.

Signed-off-by: Balazs Scheidler <balazs.scheidler@balabit.com>