nvtable-serialize: create the updated NVDynValue array separately
Instead of poking into the dynamic value table of the NVTable during the
update operation, do that in a separate memory area.  This makes it possible
to do handle updates incrementally, using the same loop.

Earlier, we had to do two stages during update: first we updated referenced
handles and SDATA entries and then with another loop the dynamic entry
table. The reason was that this update operation broke the sorted nature
of the dynamic entry table, which also broke handle based lookup.

This change creates a separate NVDynValue array where we put our updates,
then we sort and move it in if everything goes as planned.

Also makes the code easier to follow and allows a number of optimizations
in the future.

Signed-off-by: Balazs Scheidler <balazs.scheidler@balabit.com>