nvtable-serialize: change LogMessage on-disk format to support the "unset" bit
Earlier syslog-ng versions didn't have a support for the unset operation,
thus they might have set that bit to garbage. This patch changes
the deserialization code, so that based on the NVTable header, it
clears the unset bit (and the rest of the NVEntry bits as well) in case
the current NVTable is read from an earlier version.

Although this is a change in the on-disk format, this is NOT a version bump,
for the following reasons:

1) old syslog-ng writes, new one reads:

   The NVTable header bit NVT_SUPPORTS_UNSET is going to be zero, we clear
   the unset bit on all NVEntry instances. All NVEntries retain their
   value.

2) new syslog-ng writes, old one reads:

   The NVTable heeader bit NVT_SUPPORTS_UNSET is non-zero, which the
   old syslog-ng ignores. The unset bit on NVEntry instances also
   remains set, but that is also ignored by the old syslog-ng instance.

   Whenever the new syslog-ng actually unset()s a value, that'll at the
   same time set that value to the empty string, so unset() values
   will not trickle into old syslog-ng versions either.

Signed-off-by: Balazs Scheidler <balazs.scheidler@balabit.com>