groupunset: there is set/unset rewrite rules, but the groupset had no pair.
This patch implements the groupunset rewrite rule.
Previously you can unset a batch of values using the groupset with empty value
But, the empty value handling is introduced, this method does not work anymore.

So we need a rewrite rule which can do this job.

That is what this patch does.

usage example groupunset(values(".SDATA.*"));

Signed-off-by: Viktor Juhasz <viktor.juhasz@balabit.com>