logproto-server: document the layering violation of log_proto_server_get_fd()
Earlier only a note was saying that get_fd() is a layering violation. The
reason is that the presence of this function exposes that all LogProto
instances are based on a UNIX fd. This is always true as of now, but
still not nice. It would be a lot better to replace log_proto_server_get_fd()
with APIs that abstract away fds.

Signed-off-by: Balazs Scheidler <balazs.scheidler@balabit.com>